move: SessionManager 패키지 이동
util에서 application으로 이동합니다.
